Some fix

+++ Worms_SFML/Worms_SFML/camera.cpp
@@ -1,6 +1,6 @@
 #include <includes.hpp>
 
-sf::View camera(sf::FloatRect(0, 0, WINDOW_WIDTH, WINDOW_HEIGHT)); //объявление вида
+sf::View camera(sf::FloatRect(0, 0, (float)WINDOW_WIDTH, (float)WINDOW_HEIGHT)); //объявление вида
 
 void getPlayerCoordinateForView() {
 	b2Vec2 position = body[currentWorm]->GetPosition();
@@ -15,49 +15,49 @@ void getPlayerCoordinateForView() {
 		if ((position.y*SCALE < -1000 + WINDOW_HEIGHT / 2) &&
 			!(position.x*SCALE < WINDOW_WIDTH / 2) &&
 			!(position.x*SCALE > 2500 * (WINDOW_WIDTH / backgroundSprite.getLocalBounds().width))) {
-			camera.setCenter(position.x* SCALE, -1000 + WINDOW_HEIGHT / 2);
+			camera.setCenter((float)(position.x* SCALE), (float)(-1000 + WINDOW_HEIGHT / 2));
 		}
 		//Вверх и влево
 		else if ((position.y*SCALE < -1000 + WINDOW_HEIGHT / 2) &&
 			(position.x*SCALE < WINDOW_WIDTH / 2)) {
-			camera.setCenter(WINDOW_WIDTH / 2, -1000 + WINDOW_HEIGHT / 2);
+			camera.setCenter((float)(WINDOW_WIDTH / 2), (float)(-1000 + WINDOW_HEIGHT / 2));
 		}
 		//Вверх и вправо
 		else if ((position.y*SCALE < -1000 + WINDOW_HEIGHT / 2) &&
 			(position.x*SCALE > 2500 * (WINDOW_WIDTH / backgroundSprite.getLocalBounds().width))) {
-			camera.setCenter(2500 - WINDOW_WIDTH / 2, -1000 + WINDOW_HEIGHT / 2);
+			camera.setCenter((float)(2500 - WINDOW_WIDTH / 2),(float)(-1000 + WINDOW_HEIGHT / 2));
 		}
 		//Только вниз
 		else if ((position.y*SCALE > WINDOW_HEIGHT / 2) &&
 			!(position.x*SCALE < WINDOW_WIDTH / 2) &&
 			!(position.x*SCALE > 2500 * (WINDOW_WIDTH / backgroundSprite.getLocalBounds().width))) {
-			camera.setCenter(position.x*SCALE, WINDOW_HEIGHT / 2);
+			camera.setCenter((float)(position.x*SCALE), (float)(WINDOW_HEIGHT / 2));
 
 		}
 		//Вниз и влево
 		else if ((position.y*SCALE > WINDOW_HEIGHT / 2) &&
 			(position.x*SCALE < WINDOW_WIDTH / 2)) {
-			camera.setCenter(WINDOW_WIDTH / 2, WINDOW_HEIGHT / 2);
+			camera.setCenter((float)(WINDOW_WIDTH / 2), (float)(WINDOW_HEIGHT / 2));
 		}
 		//Вниз и вправо
 		else if ((position.y*SCALE > WINDOW_HEIGHT / 2) &&
 			(position.x*SCALE > 2500 * (WINDOW_WIDTH / backgroundSprite.getLocalBounds().width))) {
-			camera.setCenter(2500 - WINDOW_WIDTH / 2, WINDOW_HEIGHT / 2);
+			camera.setCenter((float)(2500 - WINDOW_WIDTH / 2), (float)(WINDOW_HEIGHT / 2));
 		}
 		//Только влево
 		else if ((position.x*SCALE < WINDOW_WIDTH / 2) &&
 			!(position.y*SCALE > WINDOW_HEIGHT / 2) &&
 			!(position.y*SCALE < -1000 + WINDOW_HEIGHT / 2)) {
-			camera.setCenter(WINDOW_WIDTH / 2, position.y*SCALE);
+			camera.setCenter((float)(WINDOW_WIDTH / 2), (float)(position.y*SCALE));
 		}
 		//Только вправо
 		else if ((position.x*SCALE > 2500 * (WINDOW_WIDTH / backgroundSprite.getLocalBounds().width)) &&
 			!(position.y*SCALE > WINDOW_HEIGHT / 2) &&
 			!(position.y*SCALE < -1000 + WINDOW_HEIGHT / 2)) {
-			camera.setCenter(2500 - WINDOW_WIDTH / 2, position.y*SCALE);
+			camera.setCenter((float)(2500 - WINDOW_WIDTH / 2),(float)( position.y*SCALE));
 		}
 }
 	//Если всё норм
 	else
-	camera.setCenter(position.x* SCALE, position.y*SCALE);
+	camera.setCenter((float)(position.x* SCALE), (float)(position.y*SCALE));
 };
